Update property-based for steering governor Prior to this commit, one of the property based tests on the steering governor was failing due to the new scaling factor from the last commit. This commit updates the property based tests for the steering module to use both the new data type and scaling factor introduced in the previous commit.

+++ platforms/kia_soul/firmware/steering/tests/property/src/tests.rs
@@ -478,7 +478,7 @@ fn check_rx_chassis_2() {
 
 /// the steering module should output lower values of torque
 /// when the vehicle is operating at higher speeds
-fn prop_check_torque_constraints(vehicle_speed: i16,
+fn prop_check_torque_constraints(vehicle_speed: f32,
                                  current_steering_angle: f32,
                                  previous_steering_angle: f32,
                                  commanded_steering_angle: f32,
@@ -541,5 +541,5 @@ fn check_torque_constraints() {
     QuickCheck::new()
         .tests(1000)
         .gen(StdGen::new(rand::thread_rng(), i16::max_value() as usize))
-        .quickcheck(prop_check_torque_constraints as fn(i16, f32, f32, f32, pid_s) -> TestResult)
+        .quickcheck(prop_check_torque_constraints as fn(f32, f32, f32, f32, pid_s) -> TestResult)
 }
